feat!: remove backward-compatibility. remove export default and named export 'authentication'
